requirements.txt
Werkzeug==0.16.0
flask
flask-bootstrap4
flask-migrate
flask-sqlalchemy
flask-testing
flask-wtf
flask_fontawesome
gensim
nltk
pandas
pycodestyle
pylint
pylint-flask
pylint_flask_sqlalchemy
pytest
pytest-cov
pytest-profiling
Flask-Session
pyvirtualdisplay
rq
redis
requests
selenium
simplejson
sklearn
spacy
wheel
